+++ examples/advanced/src/gen/clients/swr/petController/useAddPetSWR.ts
@@ -1,8 +1,8 @@
-import type { SWRMutationConfiguration } from 'swr/mutation'
-import useSWRMutation from 'swr/mutation'
 import type fetch from '../../../../axios-client.ts'
-import type { RequestConfig, ResponseConfig, ResponseErrorConfig } from '../../../../axios-client.ts'
-import type { AddPet405, AddPetMutationRequest, AddPetMutationResponse } from '../../../models/ts/petController/AddPet.ts'
+import useSWRMutation from 'swr/mutation'
+import type { RequestConfig, ResponseErrorConfig, ResponseConfig } from '../../../../axios-client.ts'
+import type { AddPetMutationRequest, AddPetMutationResponse, AddPet405 } from '../../../models/ts/petController/AddPet.ts'
+import type { SWRMutationConfiguration } from 'swr/mutation'
 import { addPet } from '../../axios/petService/addPet.ts'
 
 export const addPetMutationKeySWR = () => [{ url: '/pet' }] as const