+++ examples/advanced/src/gen2/types/LogoutUser.ts
@@ -1,11 +1,11 @@
 /**
  * @description successful operation
- */
-export type LogoutUserError = any
+*/
+export type LogoutUserError = any;
 
-export type LogoutUserQueryResponse = any
+export type LogoutUserQueryResponse = any;
 
 export type LogoutUserQuery = {
-  Response: any
-  Errors: any
-}
+    Response: any;
+    Errors: any;
+};